+++ src/vue/Details/Password/Revisions.vue
@@ -1,7 +1,10 @@
 <template>
     <ul class="revision-list">
-        <li class="revision" v-for="revision in getRevisions" :key="revision.id" @click="viewAction(revision)">
-            <img class="icon" :src="revision.icon" alt="">
+        <li class="revision" v-for="revision in getRevisions" :key="revision.id" @click="viewAction(revision)" :title="getRevisionLabel(revision)">
+            <img class="icon"
+                 :src="revision.icon"
+                 :alt="getRevisionLabel(revision)"
+                 loading="lazy">
             <span class="label">{{ revision.label }}</span>
             <span class="time">{{ getDateTime(revision.created) }}</span>
             <translate class="restore"
@@ -47,6 +50,9 @@
             },
             getDateTime(date) {
                 return Localisation.formatDateTime(date);
+            },
+            getRevisionLabel(revision) {
+                return `${revision.label} – ${this.getDateTime(revision.created)}`;
             }
         }
     };
@@ -90,6 +96,8 @@
                 grid-area     : icon;
                 border-radius : var(--border-radius);
                 align-self    : center;
+                width         : 32px;
+                height        : 32px;
             }
 
             .restore {